// apps/api/src/copart/dto/query-copart.dto.d.ts
export declare class QueryCopartDto {
    page?: number;
    limit?: number;
    search?: string;
    make?: string;
    year?: number;
    yearMin?: number;
    yearMax?: number;
    damageDescription?: string;
    saleStatus?: string;
    locationState?: string;
    minPrice?: number;
    maxPrice?: number;
    minOdometer?: number;
    maxOdometer?: number;
    hasKeys?: string;
    runsDrives?: string;
    saleTitleType?: string;
    titleCategory?: string;
    sellerCategory?: string;
    color?: string;
    cylinders?: string;
    drive?: string;
    bodyStyle?: string;
    fuelType?: string;
    transmission?: string;
    engineSizeMin?: number;
    engineSizeMax?: number;
    zip?: string;
    radius?: number;
    trim?: string;
    saleDateFrom?: number;
    saleDateTo?: number;
    sortBy?: string;
    sortOrder?: 'asc' | 'desc';
    model?: string;
    ids?: string;
    inspectableOnly?: boolean;
    runsAndDrives?: boolean;
}